requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.23
celery==3.1.23
decorator==4.0.9
Django==1.9.4
django-extensions==1.6.1
django-reversion==1.10.1
gunicorn==19.5.0
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
kombu==3.0.35
MySQL-python==1.2.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
pytz==2016.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
